requirements.txt
chromadb==0.5.5
gradio==4.44.0
litellm==1.44.28
python-dotenv==1.0.1
sentence-transformers==3.0.1
transformers==4.41.2
tokenizers==0.19.1
